requirements.txt
wikipedia
python-aiml
requests
speechrecognition
pyaudio
pyttsx3
gtts
json
sklearn
nltk
kivy
bs4